PR changes to permission checking

+++ services/hexathons/src/routes/visit.ts
@@ -8,25 +8,32 @@ import { VisitModel, Visit } from "../models/visit";
 
 export const sponsorVisitRouter = express.Router();
 
-// get visit
 sponsorVisitRouter.route("/").get(
   checkAbility("read", "Visit"),
   asyncHandler(async (req, res) => {
-    const company = await apiCall(
+    if (!req.query.hexathon) {
+      throw new BadRequestError("Hexathon is required parameter");
+    }
+    const companies = await apiCall(
       Service.USERS,
-      { method: "GET", url: `/companies/employees/${req.user?.uid}` },
+      {
+        method: "GET",
+        url: `/companies/employees/${req.user?.uid}`,
+        params: {
+          hexathon: req.query.hexathon,
+        },
+      },
       req
     );
 
-    if (!company) {
-      throw new BadRequestError("Current user not associated with a company");
+    if (companies.length === 0) {
+      throw new BadRequestError("Current user not associated with company in hexathon");
     }
 
     const filter: FilterQuery<Visit> = {};
-    filter.company = String(company.id);
-    if (req.query.hexathon) {
-      filter.hexathon = String(req.query.hexathon);
-    }
+    filter.company = String(companies[0].id);
+    filter.hexathon = String(req.query.hexathon);
+
     if (req.query.visitorId) {
       filter.visitorId = String(req.query.visitorId);
     }
@@ -39,21 +46,30 @@ sponsorVisitRouter.route("/").get(
 sponsorVisitRouter.route("/").post(
   checkAbility("create", "Visit"),
   asyncHandler(async (req, res) => {
-    // get company of current user
-    const company = await apiCall(
+    if (!req.body.hexathon) {
+      throw new BadRequestError("Hexathon is required body field");
+    }
+
+    const companies = await apiCall(
       Service.USERS,
-      { method: "GET", url: `/companies/employees/${req.user?.uid}` },
+      {
+        method: "GET",
+        url: `/companies/employees/${req.user?.uid}`,
+        params: {
+          hexathon: req.body.hexathon,
+        },
+      },
       req
     );
 
-    if (!company || !req.user) {
+    if (companies.length === 0 || !req.user) {
       throw new BadRequestError("Current user not associated with a company");
     }
 
     const visit = await VisitModel.create({
       visitorId: req.body.visitorId,
       hexathon: req.body.hexathon,
-      company: company.id,
+      company: companies[0].id,
       employee: req.user.uid,
       starred: req.body.starred,
       tags: req.body.tags,
@@ -69,24 +85,29 @@ sponsorVisitRouter.route("/:visitId").get(
   checkAbility("read", "Visit"),
   asyncHandler(async (req, res) => {
     // get company of current user
-    const company = await apiCall(
+
+    const { visitId } = req.params;
+    const visit = await VisitModel.findById(visitId);
+
+    if (!visit) {
+      throw new BadRequestError("No accessible visit associated with provided visitId");
+    }
+    console.log(visit);
+
+    const companies = await apiCall(
       Service.USERS,
-      { method: "GET", url: `/companies/employees/${req.user?.uid}` },
+      {
+        method: "GET",
+        url: `/companies/employees/${req.user?.uid}`,
+        params: {
+          hexathon: String(visit.hexathon),
+        },
+      },
       req
     );
 
-    if (!company) {
-      throw new BadRequestError("Current user not associated with a company");
-    }
-
-    const { visitId } = req.params;
-    const visit = await VisitModel.findOne({
-      _id: visitId,
-      company: company.id,
-    });
-
-    if (!visit) {
-      throw new BadRequestError("No visit associated with provided visitId");
+    if (companies.length === 0 || String(visit.company) !== companies[0].id) {
+      throw new BadRequestError("No accessible visit associated with provided visitId");
     }
 
     return res.status(200).send(visit?.toJSON());
@@ -96,21 +117,26 @@ sponsorVisitRouter.route("/:visitId").get(
 sponsorVisitRouter.route("/:visitId").put(
   checkAbility("update", "Visit"),
   asyncHandler(async (req, res) => {
-    const company = await apiCall(
+    const visit = await VisitModel.findById(req.params.visitId);
+    if (!visit) {
+      throw new BadRequestError("No accessible visit associated with provided visitId");
+    }
+    const companies = await apiCall(
       Service.USERS,
-      { method: "GET", url: `/companies/employees/${req.user?.uid}` },
+      {
+        method: "GET",
+        url: `/companies/employees/${req.user?.uid}`,
+        params: {
+          hexathon: String(visit.hexathon),
+        },
+      },
       req
     );
 
-    if (!company) {
-      throw new BadRequestError("Current user not associated with a company");
+    if (companies.length === 0 || String(visit.company) !== companies[0].id) {
+      throw new BadRequestError("No accessible visit associated with provided visitId");
     }
-
-    const newVisit = await VisitModel.findOneAndUpdate(
-      {
-        _id: req.params.visitId,
-        company: company.id,
-      },
+    const updatedVisit = await visit?.update(
       {
         starred: req.body.starred,
         tags: req.body.tags,
@@ -118,21 +144,31 @@ sponsorVisitRouter.route("/:visitId").put(
       },
       { new: true }
     );
-    return res.status(200).send(newVisit?.toJSON());
+    return res.status(200).send(updatedVisit);
   })
 );
 
 sponsorVisitRouter.route("/:visitId").delete(
   checkAbility("update", "Visit"),
   asyncHandler(async (req, res) => {
-    const company = await apiCall(
+    const visit = await VisitModel.findById(req.params.visitId);
+    if (!visit) {
+      throw new BadRequestError("No accessible visit associated with provided visitId");
+    }
+    const companies = await apiCall(
       Service.USERS,
-      { method: "GET", url: `/companies/employees/${req.user?.uid}` },
+      {
+        method: "GET",
+        url: `/companies/employees/${req.user?.uid}`,
+        params: {
+          hexathon: String(visit.hexathon),
+        },
+      },
       req
     );
 
-    if (!company) {
-      throw new BadRequestError("Current user not associated with a company");
+    if (companies.length === 0 || String(visit.company) !== companies[0].id) {
+      throw new BadRequestError("No accessible visit associated with provided visitId");
     }
 
     await VisitModel.findByIdAndDelete(req.params.visitId);
